Make App container clearer on counter example
* Add CounterContainer that will export the connected component
* Add some docstrings to make connect params clearer on first read.
* App container will render the CounterContainer

The motivation was trying to read this code but failing on first try, because I couldn't understand what was happening here without writing this myself.
