+++ examples/counter/components/Counter.js
@@ -4,9 +4,8 @@ class Counter extends Component {
   render() {
     const { increment, incrementIfOdd, incrementAsync, decrement, counter } = this.props
     return (
-      <p>
+      <span>
         Clicked: {counter} times
-        {' '}
         <button onClick={increment}>+</button>
         {' '}
         <button onClick={decrement}>-</button>
@@ -14,7 +13,7 @@ class Counter extends Component {
         <button onClick={incrementIfOdd}>Increment if odd</button>
         {' '}
         <button onClick={() => incrementAsync()}>Increment async</button>
-      </p>
+      </span>
     )
   }
 }
